+++ src/Libraries/SmartStore.Core/Domain/Customers/SystemCustomerAttributeNames.cs
@@ -20,6 +20,7 @@ namespace SmartStore.Core.Domain.Customers
 		public static string VatNumber { get { return "VatNumber"; } }
 		public static string VatNumberStatusId { get { return "VatNumberStatusId"; } }
 		public static string TimeZoneId { get { return "TimeZoneId"; } }
+        public static string CustomerNumber { get { return "CustomerNumber"; } }
 
         //Other attributes
 		public static string DiscountCouponCode { get { return "DiscountCouponCode"; } }