+++ src/Libraries/SmartStore.Data/Migrations/201506181858349_AclRecordCustomerRole.cs
@@ -131,6 +131,18 @@ namespace SmartStore.Data.Migrations
                 "Zeige Herstellerbilder auf der Homepage",
                 "Specifies whether manufacturers will be displayed as images or textual links on the homepage.",
                 "Bestimmt ob Hersteller auf der Homepage als Bilder oder textuelle Links angezeigt werden.");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.CustomerUser.CustomerNumberEnabled",
+                "Customers can enter a customer number",
+                "Kunden können Kundennummer hinterlegen",
+                "Specifies whether customers can enter a customer number.",
+                "Bestimmt ob Kunden eine Kundennummer hinterlegen können.");
+
+            builder.AddOrUpdate("Account.Fields.CustomerNumber",
+                "Customer number",
+                "Kundennummer");
+            
+
         }
     }
 }